Reworked CData interface and implementation to be more consistent and optimized. Enhanced TBuffer interface. Updated CBitmap to use a TBuffer for pixel data instead of CData. Updated some byte counts to UInt64. Implemented storage for lots of modules. Enhancements and bug fixes in CCoreAudioDecodeAudioCodec. Enhancements to CMediaPacketSource. Added convenience constructor to SMedia::PacketData. Added iterators to TLockingDictionary. Added and renamed methods in CDataSource. Reworked CAudioCodec::FrameSource. Reworked CAudioFrames implementation to remove dependency on CData. Lots of updates to various modules to use the new CBitmap and CData interfaces.

+++ Source/Add On - Apple/CString-Apple.cpp
@@ -58,7 +58,7 @@ CString::CString(const OSStringVar(initialString)) : CHashable()
 }
 
 //----------------------------------------------------------------------------------------------------------------------
-CString::CString(const void* ptr, UInt32 byteCount, Encoding encoding)
+CString::CString(const void* ptr, UInt64 byteCount, Encoding encoding)
 //----------------------------------------------------------------------------------------------------------------------
 {
 	// Parameter check
@@ -499,14 +499,14 @@ TBuffer<char> CString::getUTF8Chars() const
 	// Check length
 	if (length > 0) {
 		// Setup
-		TBuffer<char>	buffer((UInt32) length * 4);
+		TBuffer<char>	buffer((UInt64) (length * 4));
 
 		// Convert
 		CFIndex	count =
 						::CFStringGetBytes(mStringRef, CFRangeMake(0, length), kCFStringEncodingUTF8, 0, false,
 								(UInt8*) *buffer, length * 4, nil);
 
-		return TBuffer<char>(buffer, (UInt32) count);
+		return TBuffer<char>(buffer, (UInt64) count);
 	} else
 		// Empty
 		return TBuffer<char>(0);
